+++ src/deprecated/Api.js
@@ -10,35 +10,38 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const express = require("express");
-const { DEFAULT_BASE_PATH } = require("@mocks-server/admin-api-paths");
+const {
+  DEFAULT_BASE_PATH,
+  BEHAVIORS,
+  FIXTURES,
+  LEGACY,
+} = require("@mocks-server/admin-api-paths");
 
 const Behaviors = require("./Behaviors");
-const Settings = require("./Settings");
+const Fixtures = require("./Fixtures");
 
-const { DEPRECATED_API_PATH, PLUGIN_NAME } = require("../constants");
+const { PLUGIN_NAME } = require("../support/constants");
 
 // TODO, deprecate mocks router
 
 class Api {
-  constructor(core) {
+  constructor(core, { addAlert }) {
     this._core = core;
     this._tracer = core.tracer;
+    this._addAlert = addAlert;
   }
 
   init() {
     const behaviorsRouter = new Behaviors(this._core).router;
+    const fixturesRouter = new Fixtures(this._core).router;
     this._router = express.Router();
     this._router.use((req, res, next) => {
-      this._core.tracer.deprecationWarn(
-        `"${DEPRECATED_API_PATH}" ${PLUGIN_NAME} path`,
-        DEFAULT_BASE_PATH
-      );
+      this._addAlert("legacy", `Detected usage of deprecated api path "${LEGACY}"`);
+      this._core.tracer.deprecationWarn(`"${LEGACY}" ${PLUGIN_NAME} path`, DEFAULT_BASE_PATH);
       next();
     });
-
-    this._router.use("/features", behaviorsRouter);
-    this._router.use("/behaviors", behaviorsRouter);
-    this._router.use("/settings", new Settings(this._core.settings, this._tracer).router);
+    this._router.use(FIXTURES, fixturesRouter);
+    this._router.use(BEHAVIORS, behaviorsRouter);
   }
 
   get router() {